# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Museum Ticketing & Booking Platform API
  version: 0.2.0
  description: Bilingual (EN/AM) online booking, payment, and gate check-in for the
    Zoological Natural History Museum's digital ticketing track.
paths:
...
  /api/v1/bookings/:
    get:
      operationId: listBookings
      description: |-
        GET /bookings (Staff only) and POST /bookings (Visitor only,
        FR-BOOK-001/FR-BOOK-003).
      parameters:
      - in: query
        name: bookingType
        schema:
          type: string
          enum:
          - group
          - individual
        description: Only return individual or group bookings.
      - in: query
        name: flagged
        schema:
          type: boolean
        description: When `true`, only return bookings flagged for a Museum Manager's
          mismatch review.
      - in: query
        name: status
        schema:
          type: string
          enum:
          - awaiting_payment
          - cancelled
          - pending
          - refunded
          - visited
        description: Only return bookings in this lifecycle status.
      - in: query
        name: visitDate
        schema:
          type: string
          format: date
        description: Only return bookings for this visit date, ISO-8601 `YYYY-MM-DD`.
      tags:
      - v1
      security:
      - TokenVersionAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: ''
    post:
      operationId: createBooking
      description: |-
        GET /bookings (Staff only) and POST /bookings (Visitor only,
        FR-BOOK-001/FR-BOOK-003).
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BookingCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BookingCreate'
        required: true
      security:
      - TokenVersionAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: ''
  /api/v1/bookings/{id}/:
    get:
      operationId: getBooking
      description: |-
        GET /bookings/{id} -- the owning Visitor or any Staff member.

        Also handles PATCH /bookings/{id} -- the owning Visitor only, editing
        their own booking's ticket mix and/or visit date while it's still
        `AwaitingPayment` (see `BookingUpdateSerializer`/`services.
        update_awaiting_payment_booking`). Kept on the same resource/URL as
        the GET above rather than a separate endpoint, since it's the same
        `Booking` -- just a different HTTP method and a narrower audience.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - TokenVersionAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: ''
    patch:
      operationId: updateBooking
      description: |-
        GET /bookings/{id} -- the owning Visitor or any Staff member.

        Also handles PATCH /bookings/{id} -- the owning Visitor only, editing
        their own booking's ticket mix and/or visit date while it's still
        `AwaitingPayment` (see `BookingUpdateSerializer`/`services.
        update_awaiting_payment_booking`). Kept on the same resource/URL as
        the GET above rather than a separate endpoint, since it's the same
        `Booking` -- just a different HTTP method and a narrower audience.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedBookingUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedBookingUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedBookingUpdate'
      security:
      - TokenVersionAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: ''
...
  /api/v1/users/me/bookings/:
    get:
      operationId: listMyBookings
      description: |-
        GET /users/me/bookings -- FR-ACC-004. Mounted at the root URLconf
        (config/urls.py), not here, mirroring CurrentUserView's placement --
        Document 04 puts this under `/users/`, not `/bookings/`.
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      tags:
      - v1
      security:
      - TokenVersionAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBookingList'
          description: ''
components:
  schemas:
...
    Booking:
      type: object
      description: |-
        `Booking` (Document 04). Read-only -- every mutation goes through
        the Create/Cancel/Reschedule/CategoryCorrection serializers below and
        services.py, never through this serializer directly.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        reference:
          type: string
          description: The 8-character booking reference, drawn from the alphabet
            `ABCDEFGHJKLMNPQRSTUVWXYZ23456789` (ambiguous `0/O/1/I` excluded). The
            Visitor presents it at the gate, typed or scanned.
          minLength: 8
          maxLength: 8
          pattern: ^[A-HJ-NP-Z2-9]{8}$
          example: K7M2QP4A
          readOnly: true
        visitorId:
          type: string
          format: uuid
          readOnly: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookingItem'
          readOnly: true
        visitDate:
          type: string
          format: date
          readOnly: true
        bookingType:
          type: string
          readOnly: true
        groupName:
          type: string
          readOnly: true
          nullable: true
        groupContactPhone:
          type: string
          readOnly: true
          nullable: true
        groupTin:
          type: string
          readOnly: true
          nullable: true
        institutionId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        visitorName:
          type: string
          readOnly: true
        visitorEmail:
          type: string
          readOnly: true
        visitorPhone:
          type: string
          readOnly: true
          nullable: true
        bookedQuantity:
          type: integer
          readOnly: true
        attendedQuantity:
          type: integer
          readOnly: true
          nullable: true
        status:
          allOf:
          - $ref: '#/components/schemas/BookingStatusEnum'
          readOnly: true
        rescheduledCount:
          type: integer
          readOnly: true
        categoryCorrectedAt:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        flaggedMismatchAt:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        flaggedMismatchByUserId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        flaggedMismatchNote:
          type: string
          readOnly: true
          nullable: true
        noticeSentAt:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        checkoutUrl:
          type: string
          readOnly: true
          nullable: true
        receiptUrl:
          type: string
          readOnly: true
          nullable: true
        ifmisDocumentNo:
          type: string
          readOnly: true
          nullable: true
        ifmisVoucherReference:
          type: string
          readOnly: true
          nullable: true
        reconciliationId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        totalAmountEtb:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
      required:
      - attendedQuantity
      - bookedQuantity
      - bookingType
      - categoryCorrectedAt
      - checkoutUrl
      - createdAt
      - flaggedMismatchAt
      - flaggedMismatchByUserId
      - flaggedMismatchNote
      - groupContactPhone
      - groupName
      - groupTin
      - id
      - ifmisDocumentNo
      - ifmisVoucherReference
      - institutionId
      - items
      - noticeSentAt
      - receiptUrl
      - reconciliationId
      - reference
      - rescheduledCount
      - status
      - totalAmountEtb
      - visitDate
      - visitorEmail
      - visitorId
      - visitorName
      - visitorPhone
...
    BookingCreate:
      type: object
      description: |-
        `BookingCreateRequest` -- FR-BOOK-001 (individual), FR-BOOK-003
        (group). A plain Serializer, not a ModelSerializer: `visitor` is
        supplied by the view from `request.user`, never from the request
        body, and `items`/`bookingType` need their own field names mapped
        onto services.create_booking's kwargs rather than a 1:1 model field
        mapping.

        `items` replaces the old single `categoryId`/`quantity` pair -- a
        booking can now mix categories (one Adult plus two Student tickets in
        the same checkout) instead of being limited to one category with a
        plain headcount. `min_length=1`: every booking still needs at least
        one category, same as before.
      properties:
        visitDate:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookingCreateItem'
        bookingType:
          $ref: '#/components/schemas/BookingTypeEnum'
        groupName:
          type: string
          nullable: true
        groupContactPhone:
          type: string
          nullable: true
        groupTin:
          type: string
          nullable: true
      required:
      - bookingType
      - items
      - visitDate
...
    BookingStatusEnum:
      enum:
      - awaiting_payment
      - pending
      - visited
      - cancelled
      - refunded
      type: string
      description: |-
        * `awaiting_payment` - Awaiting payment
        * `pending` - Pending
        * `visited` - Visited
        * `cancelled` - Cancelled
        * `refunded` - Refunded
...
    BookingTypeEnum:
      enum:
      - individual
      - group
      type: string
      description: |-
        * `individual` - Individual
        * `group` - Group
...
    PaginatedBookingList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Booking'
        meta:
          type: object
          properties:
            limit:
              type: integer
            offset:
              type: integer
            total:
              type: integer
          required:
          - limit
          - offset
          - total
      required:
      - data
      - meta
...
    PatchedBookingUpdate:
      type: object
      description: |-
        `BookingUpdateRequest` -- Visitor-only edit of their own still-
        unpaid (`AwaitingPayment`) booking (see
        `services.update_awaiting_payment_booking`). Both `items` and
        `visitDate` are optional here, unlike `BookingCreateSerializer`: a
        Visitor may only be fixing one of the two. There is no
        `bookingType`/`groupName`/`groupTin` here -- Document 02 has no
        notion of changing what *kind* of booking this is after the fact,
        only its ticket mix and date.
      properties:
        visitDate:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookingUpdateItem'
...
  securitySchemes:
    TokenVersionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT